+++ datacube_zarr/utils/uris.py
@@ -1,5 +1,8 @@
-from typing import Optional, Tuple
+from typing import MutableMapping, Optional, Tuple
 from urllib.parse import urlparse
+
+import s3fs
+import zarr
 
 
 def uri_split(uri: str) -> Tuple[str, str, str]:
@@ -34,3 +37,29 @@ def uri_join(protocol: str, root: str, group: Optional[str] = None) -> str:
     """
     uri = f"{protocol}://{root}" + (f"#{group}" if group else "")
     return uri
+
+
+def uri_to_store_and_group(uri: str) -> Tuple[MutableMapping, str]:
+    """Convert a '<protocol>://<path>#<group>' sting to a zarr storage class and group."""
+
+    # With new s3fs release we can use zarr.FSStore directly or via `normalize_store_args`
+    #
+    # Something like:
+    #
+    # store_uri, group = uri.split("#", 1) if "#" in uri else (uri, '')
+    # storage_options = {"normalize_keys": False}
+    # store = zarr.creation.normalize_store_arg(
+    #    store_uri, clobber=True, storage_options=storage_options
+    # )
+
+    protocol, root, group = uri_split(uri)
+    if protocol == 's3':
+        s3 = s3fs.S3FileSystem()
+        s3.invalidate_cache()
+        store = s3.get_mapper(root=root, check=False)
+    elif protocol == 'file':
+        store = zarr.DirectoryStore(root)
+    else:
+        raise ValueError(f'Protocol not known: {protocol}')
+
+    return store, group