+++ tests/test_zarr_io.py
@@ -6,24 +6,39 @@ import xarray as xr
 
 from datacube_zarr.zarr_io import ZarrIO, replace_dataset_dim
 
-from .utils import _check_zarr_files, _load_dataset, _save_dataarray, _save_dataset
+from .utils import _check_zarr_files, _load_dataset, _save_dataset
 
 
-def test_consolidated_metadata_exists(tmp_3d_zarr):
-    """Test that s3fs cache is working."""
-    root = ZarrIO().get_root(tmp_3d_zarr)
-    is_consolidated = ".zmetadata" in root
-    assert is_consolidated
+# Remove this when updating to latest s3fs.
+# This bug stops us from using `zarr.FSStore`.
+@pytest.mark.xfail(reason="https://github.com/dask/s3fs/issues/475")
+def test_s3fs(tmp_s3path):
+    base = tmp_s3path / "base"
+    root = str(base)[1:]
+    import s3fs
+
+    fs = s3fs.S3FileSystem()
+    fs.touch(f"{root}/file.abc")
+    fs.touch(f"{root}/sub/file2.abc")
+    fs.find(root)
+    ls_res = fs.ls(root)
+    assert f"{root}/file.abc" in ls_res
+    assert f"{root}/sub" in ls_res
 
 
-@pytest.mark.parametrize("save_fn", [_save_dataarray, _save_dataset])
-def test_save_dataarray(tmp_storage_path, chunks, data, save_fn):
-    '''Test ZarrIO.save_dataarray to save and load for a single DataArray.'''
+@pytest.mark.parametrize(
+    "group",
+    [
+        "dataset1",
+        "Dataset1",
+    ],  # pytest.param("DataSet1", marks=pytest.mark.xfail(reason="uppercase"))],
+)
+def test_save_dataset(tmp_storage_path, chunks, data, group):
+    '''Test ZarrIO.save_dataset to save and load for a single DataArray.'''
     root = tmp_storage_path / "data.zarr"
-    group = "dataset1"
     name = 'array1'
     uri = f"{root.as_uri()}#{group}"
-    save_fn(data, uri, name, chunks['input'])
+    _save_dataset(data, uri, name, chunks['input'])
 
     # Check filesystem or s3 structure and some metadata
     _check_zarr_files(data, root, group, name, chunks)
@@ -33,9 +48,8 @@ def test_save_dataarray(tmp_storage_path, chunks, data, save_fn):
     assert np.array_equal(data, ds[name].values)
 
 
-@pytest.mark.parametrize("save_fn", [_save_dataarray, _save_dataset])
-def test_save_dataarrays(tmp_storage_path, chunks, data, save_fn):
-    '''Test ZarrIO.save_dataarray to save and load multiple DataArrays.'''
+def test_save_datasets(tmp_storage_path, chunks, data):
+    '''Test ZarrIO.save_dataset to save and load multiple DataArrays.'''
     datasets = []
     root = tmp_storage_path / "data.zarr"
     for set_no in range(1, 3):
@@ -44,7 +58,7 @@ def test_save_dataarrays(tmp_storage_path, chunks, data, save_fn):
         name = f'array{set_no}'
         ds = data.copy() * set_no  # Make each dataset a bit different for testing
         datasets.append((uri, name, ds))
-        save_fn(ds, uri, name, chunks['input'])
+        _save_dataset(ds, uri, name, chunks['input'])
 
     assert len(datasets) == 2
 
@@ -55,29 +69,17 @@ def test_save_dataarrays(tmp_storage_path, chunks, data, save_fn):
 
 
 def test_print_tree(tmp_storage_path, fixed_chunks, data):
-    '''Test zarr print data tree with a mix of Datasets and DataArrays co-existing.'''
+    '''Test zarr print data tree with a mix of Datasets.'''
     root = tmp_storage_path / "data.zarr"
     root_uri = root.as_uri()
     zio = ZarrIO()
-    zio.save_dataarray(
-        uri=f'{root_uri}#dataset1',
-        dataarray=data,
-        name='array1',
-        chunks=fixed_chunks['input'],
-    )
-    zio.save_dataarray(
-        uri=f'{root_uri}#dataset2',
-        dataarray=data,
-        name='array2',
-        chunks=fixed_chunks['input'],
-    )
     zio.save_dataset(
-        uri=f'{root_uri}#dataset3',
+        uri=f'{root_uri}#dataset1',
         dataset=data.to_dataset(name='array1'),
         chunks=fixed_chunks['input'],
     )
     zio.save_dataset(
-        uri=f'{root_uri}#dataset4',
+        uri=f'{root_uri}#dataset2',
         dataset=data.to_dataset(name='array2'),
         chunks=fixed_chunks['input'],
     )
@@ -85,11 +87,7 @@ def test_print_tree(tmp_storage_path, fixed_chunks, data):
     expected = '''/
  ├── dataset1
  │   └── array1 (1300, 1300) float64
- ├── dataset2
- │   └── array2 (1300, 1300) float64
- ├── dataset3
- │   └── array1 (1300, 1300) float64
- └── dataset4
+ └── dataset2
      └── array2 (1300, 1300) float64'''
     assert actual == expected
 
@@ -97,6 +95,8 @@ def test_print_tree(tmp_storage_path, fixed_chunks, data):
 def test_clean_store(tmp_storage_path, fixed_chunks, data):
     '''Test cleaning of zarr store.'''
     root = tmp_storage_path / "data.zarr"
+    assert not root.exists()
+
     root_uri = root.as_uri()
     zio = ZarrIO()
     zio.save_dataset(
@@ -104,6 +104,7 @@ def test_clean_store(tmp_storage_path, fixed_chunks, data):
         dataset=data.to_dataset(name='array1'),
         chunks=fixed_chunks['input'],
     )
+    assert root.exists()
     assert (
         str(zio.print_tree(root_uri))
         == '''/
@@ -112,10 +113,11 @@ def test_clean_store(tmp_storage_path, fixed_chunks, data):
     )
     # Clean and store something else
     zio.clean_store(root_uri)
-    zio.save_dataarray(
+    assert not root.exists()
+
+    zio.save_dataset(
         uri=f'{root_uri}#dataset2',
-        dataarray=data,
-        name='array2',
+        dataset=data.to_dataset(name='array2'),
         chunks=fixed_chunks['input'],
     )
     assert (
@@ -131,7 +133,7 @@ def test_invalid_protocol():
     with pytest.raises(ValueError) as excinfo:
         zio = ZarrIO()
         zio.get_root('xxx://root')
-    assert str(excinfo.value) == 'unknown protocol: xxx'
+    assert str(excinfo.value) == "Protocol not known: xxx"
 
 
 def test_invalid_mode(example_uri, fixed_chunks, data):
@@ -141,19 +143,6 @@ def test_invalid_mode(example_uri, fixed_chunks, data):
         zio.save_dataset(
             example_uri,
             dataset=data.to_dataset(name='array1'),
-            chunks=fixed_chunks['input'],
-            mode='xxx',
-        )
-    assert (
-        str(excinfo.value)
-        == f"Only the following modes are supported {ZarrIO.WRITE_MODES}"
-    )
-
-    with pytest.raises(ValueError) as excinfo:
-        zio.save_dataarray(
-            example_uri,
-            dataarray=data,
-            name='array1',
             chunks=fixed_chunks['input'],
             mode='xxx',
         )
@@ -173,7 +162,7 @@ def test_overwrite_dataset(example_uri, fixed_chunks, data):
         zio.save_dataset(
             uri=example_uri, dataset=dataset, chunks=fixed_chunks['input'], mode='w'
         )
-    ds = zio.load_dataset(uri=example_uri)
+    ds = zio.open_dataset(uri=example_uri).load()
     np.array_equal(dataset[name], ds[name].values)
 
 
@@ -228,19 +217,19 @@ def test_save_datasets_nested_zarr(tmp_storage_path, data):
 def test_rename_dataset_dim_nocoords(example_uri, data):
     """Test rename dimension in zarr dataset with no coords."""
     _save_dataset(data, example_uri, "data")
-    ds_a = ZarrIO().load_dataset(example_uri)
+    ds_a = _load_dataset(example_uri)
     rename_dict = {"dim_0": "x", "dim_1": "y"}
     for old, new in rename_dict.items():
         replace_dataset_dim(example_uri, old, new)
 
-    ds_b = ZarrIO().load_dataset(example_uri)
+    ds_b = _load_dataset(example_uri)
     assert ds_a.rename(rename_dict).equals(ds_b)
 
 
 def test_replace_dataset_dim_nocoords(example_uri, data):
     """Test replace dimension in zarr dataset with no coords."""
     _save_dataset(data, example_uri, "data")
-    ds_a = ZarrIO().load_dataset(example_uri)
+    ds_a = _load_dataset(example_uri)
     with pytest.raises(ValueError):
         new_dim = xr.IndexVariable("x", ds_a["dim_0"].data)
         replace_dataset_dim(example_uri, "dim_0", new_dim)
@@ -248,21 +237,21 @@ def test_replace_dataset_dim_nocoords(example_uri, data):
 
 def test_rename_dataset_dim(tmp_3d_zarr):
     """Test rename dimension in zarr dataset."""
-    ds_a = ZarrIO().load_dataset(tmp_3d_zarr)
+    ds_a = _load_dataset(tmp_3d_zarr)
     replace_dataset_dim(tmp_3d_zarr, "band", "abc")
-    ds_b = ZarrIO().load_dataset(tmp_3d_zarr)
+    ds_b = _load_dataset(tmp_3d_zarr)
     assert ds_a.rename({"band": "abc"}).equals(ds_b)
 
 
 @pytest.mark.parametrize("dtype", [np.int64, np.float64])
 def test_replace_dataset_dim(tmp_3d_zarr, dtype):
     """Test replace dimension in zarr dataset."""
-    ds_a = ZarrIO().load_dataset(tmp_3d_zarr)
+    ds_a = _load_dataset(tmp_3d_zarr)
     dim_len = len(ds_a["band"])
     new_name = "lambda"
     new_dim = xr.IndexVariable(new_name, np.linspace(123, 456, dim_len).astype(dtype))
     replace_dataset_dim(tmp_3d_zarr, "band", new_dim)
-    ds_b = ZarrIO().load_dataset(tmp_3d_zarr)
+    ds_b = _load_dataset(tmp_3d_zarr)
     assert ds_b.variables[new_name].equals(new_dim)
     ds_a_new = ds_a.rename({"band": new_name}).assign_coords({new_name: new_dim})
     assert ds_a_new.equals(ds_b)
@@ -279,7 +268,7 @@ def test_replace_dataset_dim_badname(tmp_3d_zarr, old, new):
 
 def test_replace_dataset_dim_wronglen(tmp_3d_zarr):
     """Test replacing zarr dimension with bad length."""
-    ds_a = ZarrIO().load_dataset(tmp_3d_zarr)
+    ds_a = _load_dataset(tmp_3d_zarr)
     new_dim = xr.IndexVariable("band2", np.append(ds_a["band"].data, [999]))
     with pytest.raises(ValueError):
         replace_dataset_dim(tmp_3d_zarr, "band", new_dim)